Update BellaandCharlotteWebsiteCFG.html
I uploaded the first draft for the HTML code for the Home Page of our Website. This is just a start and still needs your pictures and text for your 'About' section. I am currently trying out some CSS ideas and plan on working on the HTML for the other pages on the website.