Add files via upload
This is the CSS I have been working on for the homepage. The CSS I have labelled 'Smart Sense CSS.' Let me know what you like and what you don't like. There are still adjustments that need to be made and we can use your pictures instead if you prefer.